requirements.txt
fastapi
uvicorn
python-multipart
requests
supabase
python-dotenv
gradio_client